fix: release session lock early for spellcheck requests
Spellcheck requests block on the LanguageTool backend for many seconds
while only reading the session for authentication. The PDO session
handler holds a row-level lock on the session row until the response,
so a slow spellcheck request serialized every concurrent request from
the same session, surfacing as 1205 lock wait timeout exceeded.

Generalize the former tus-only subscriber to release the session lock
early for both tus uploads and spellcheck endpoints.
